Word creation taken out of DNSPre and Post
